basic SIRs implementation
I've done an initial implementation of the SIRs model. For S -> I, we look at the 4 closest neighbours for a cell and if any of them are infected then that cell will transform with probability p1. For I->R, a cell will transform with probability p2, regardless of neighbours. For R -> S, a cell will transform with probability p3, regardless of neighbours. Implemented a matplotlib animation but a more extensive GUI will be needed for the final product. This will mean making the code more efficient. Next steps will include tracking average rates of infection, plotting those graphs and making the model more complex.