Implemented the base iterations
The basic functionality of the simulation is built in this commit.

Multiple variants can propagate on the network in a SIR model and no two nodes can be infected with multiple variants at the same time.

The code needs to be made more efficient and readable and I need to implement a 'relation' factor such that nodes recovered from one disease have a chance of being fully immune