Added Binary Trees to represent Variants
Binary trees to represent variants means that we get a more natural (and random) evolution of variants. Each disease variant is assigned a random number from 0 to 1000 and then with disease A being the root node in the tree, the other variants are inserted using the classic rules of binary trees. The difference between the numbers associated with each variant give a "strength" factor for how each disease is related.