Direct Method Gillespie Time
Changed the way time is calculated such that it follows a more accurate "Direct Method" of the Gillespie algorithm. The probabilities behind the next chosen infection (or recovery) still say the same but the time jump is based on an exponential distribution. The chosen actions, their related probabilities and the chosen next time based off of that chosen action (the First Reaction approach of Gillespie) is mathematically equivalent to this implementation